@@ -3,6 +3,7 @@
   "version": "0.12.1",
   "description": "GenLayer Command Line Tool",
   "main": "src/index.ts",
+  "type": "module",
   "bin": {
     "genlayer": "./dist/index.js"
   },
@@ -34,7 +35,7 @@
   },
   "homepage": "https://github.com/yeagerai/genlayer-cli#readme",
   "devDependencies": {
-    "@release-it/conventional-changelog": "^8.0.1",
+    "@release-it/conventional-changelog": "^10.0.0",
     "@types/dockerode": "^3.3.31",
     "@types/inquirer": "^9.0.7",
     "@types/node": "^22.0.0",
@@ -44,26 +45,28 @@
     "@typescript-eslint/parser": "^8.0.0",
     "@vitest/coverage-v8": "^2.1.4",
     "cross-env": "^7.0.3",
-    "esbuild": "^0.24.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
+    "esbuild": "^0.25.0",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^10.0.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
-    "jsdom": "^25.0.1",
+    "jsdom": "^26.0.0",
     "prettier": "^3.2.5",
-    "release-it": "^17.2.0",
+    "release-it": "^18.0.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5"
   },
   "dependencies": {
-    "commander": "^12.0.0",
+    "chalk": "^5.4.1",
+    "commander": "^13.1.0",
     "dockerode": "^4.0.2",
     "dotenv": "^16.4.5",
     "ethers": "^6.13.4",
     "genlayer-js": "^0.6.0",
-    "inquirer": "^9.3.7",
+    "inquirer": "^12.0.0",
     "node-fetch": "^3.0.0",
     "open": "^10.1.0",
+    "ora": "^8.2.0",
     "update-check": "^1.5.4",
     "uuid": "^11.0.0",
     "viem": "^2.21.54",
