@@ -1,16 +1,20 @@
 #!/usr/bin/env node
+/* eslint-disable no-undef -- Allow process and console in ignored file */
 
-const fs = require('fs');
-const path = require('path');
+import { existsSync, copyFileSync } from 'fs';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
 
-const envExamplePath = path.resolve(__dirname, '../.env.example');
-const envPath = path.resolve(__dirname, '../.env');
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const envExamplePath = resolve(__dirname, '../.env.example');
+const envPath = resolve(__dirname, '../.env');
 
 try {
-  if (fs.existsSync(envPath)) {
+  if (existsSync(envPath)) {
     console.log(`⚠️  .env file already exists. Skipping creation.`);
   } else {
-    fs.copyFileSync(envExamplePath, envPath);
+    copyFileSync(envExamplePath, envPath);
     console.log(`✅  .env file created successfully from .env.example.`);
   }
 } catch (error) {
