@@ -1,8 +1,8 @@
 import { Command } from "commander";
 
 import simulatorService from "../../lib/services/simulator";
-import { initAction, InitActionOptions } from "./init";
-import { startAction, StartActionOptions } from "./start";
+import { InitAction, InitActionOptions } from "./init";
+import { StartAction, StartActionOptions } from "./start";
 import {localnetCompatibleVersion} from "../../lib/config/simulator";
 import {StopAction} from "./stop";
 
@@ -14,7 +14,10 @@ export function initializeGeneralCommands(program: Command) {
     .option("--headless", "Headless mode", false)
     .option("--reset-db", "Reset Database", false)
     .option("--localnet-version <localnetVersion>", "Select a specific localnet version", localnetCompatibleVersion)
-    .action((options: InitActionOptions) => initAction(options, simulatorService));
+    .action(async (options: InitActionOptions) => {
+      const initAction = new InitAction();
+      await initAction.execute(options)
+    });
 
   program
     .command("up")
@@ -23,7 +26,10 @@ export function initializeGeneralCommands(program: Command) {
     .option("--numValidators <numValidators>", "Number of validators", "5")
     .option("--headless", "Headless mode", false)
     .option("--reset-db", "Reset Database", false)
-    .action((options: StartActionOptions) => startAction(options, simulatorService));
+    .action(async (options: StartActionOptions) => {
+      const startAction = new StartAction();
+      await startAction.execute(options);
+    });
 
   program
     .command("stop")
