@@ -1,104 +1,103 @@
 import inquirer from "inquirer";
-
-import {ISimulatorService} from "../../lib/interfaces/ISimulatorService";
+import { ISimulatorService } from "../../lib/interfaces/ISimulatorService";
+import { DistinctQuestion } from "inquirer";
+import { BaseAction } from "../../lib/actions/BaseAction";
+import { SimulatorService } from "../../lib/services/simulator";
 
 export interface StartActionOptions {
   resetValidators: boolean;
   numValidators: number;
   headless: boolean;
-  resetDb: boolean
+  resetDb: boolean;
 }
 
-export async function startAction(options: StartActionOptions, simulatorService: ISimulatorService) {
-  const {resetValidators, numValidators, headless, resetDb} = options;
+export class StartAction extends BaseAction {
+  private simulatorService: ISimulatorService;
 
-  simulatorService.setComposeOptions(headless);
-
-  await simulatorService.checkCliVersion();
-
-  const restartValidatorsHintText = resetValidators
-    ? `creating new ${numValidators} random validators`
-    : "keeping the existing validators";
-
-  console.log(`Starting GenLayer simulator ${restartValidatorsHintText}`);
-
-  try {
-    await simulatorService.runSimulator();
-  } catch (error) {
-    console.error(error);
-    return;
+  constructor() {
+    super();
+    this.simulatorService = new SimulatorService();
   }
 
-  try {
-    const {initialized, errorCode, errorMessage} = await simulatorService.waitForSimulatorToBeReady();
-    if (!initialized && errorCode === "ERROR") {
-      console.log(errorMessage);
-      console.error("Unable to initialize the GenLayer simulator. Please try again.");
-      return;
-    }
-    if (!initialized && errorCode === "TIMEOUT") {
-      console.error(
-        "The simulator is taking too long to initialize. Please try again after the simulator is ready.",
-      );
-      return;
-    }
-    console.log("Simulator is running!");
-  } catch (error) {
-    console.error(error);
-    return;
-  }
+  async execute(options: StartActionOptions) {
+    const { resetValidators, numValidators, headless, resetDb } = options;
 
-  if(resetDb){
-    await simulatorService.cleanDatabase()
-  }
+    this.simulatorService.setComposeOptions(headless);
+    this.startSpinner("Checking CLI version...");
+    await this.simulatorService.checkCliVersion();
 
-  if (resetValidators) {
-    // Initializing validators
-    console.log("Initializing validators...");
+    const restartValidatorsHintText = resetValidators
+      ? `creating new ${numValidators} random validators`
+      : "keeping the existing validators";
+    this.setSpinnerText(`Starting GenLayer Localnet (${restartValidatorsHintText})...`);
+
     try {
-      //remove all validators
-      await simulatorService.deleteAllValidators();
-      const questions = [
-        {
-          type: "checkbox",
-          name: "selectedLlmProviders",
-          message: "Select which LLM providers do you want to use:",
-          choices: simulatorService.getAiProvidersOptions(false),
-          validate: function (answer: string[]) {
-            if (answer.length < 1) {
-              return "You must choose at least one option.";
-            }
-            return true;
-          },
-        },
-      ];
-
-      // Since ollama runs locally we can run it here and then look for the other providers
-      const llmProvidersAnswer = await inquirer.prompt(questions);
-
-      // create random validators
-      await simulatorService.createRandomValidators(
-        Number(options.numValidators),
-        llmProvidersAnswer.selectedLlmProviders,
-      );
+      await this.simulatorService.runSimulator();
     } catch (error) {
-      console.error("Unable to initialize the validators.");
-      console.error(error);
+      this.failSpinner("Error starting the simulator", error);
       return;
     }
-    console.log("New random validators successfully created...");
-  }
 
-  // Simulator ready
-  let successMessage = "GenLayer simulator initialized successfully! "
-  successMessage += headless ? '' :  `Go to ${simulatorService.getFrontendUrl()} in your browser to access it.`;
-  console.log(successMessage);
-  try {
-    if(!headless) {
-      await simulatorService.openFrontend();
+    try {
+      this.setSpinnerText("Waiting for the simulator to be ready...");
+      const { initialized, errorCode, errorMessage } = await this.simulatorService.waitForSimulatorToBeReady();
+
+      if (!initialized) {
+        if (errorCode === "ERROR") {
+          this.failSpinner("Unable to initialize the GenLayer simulator.", errorMessage);
+          return;
+        }
+        if (errorCode === "TIMEOUT") {
+          this.failSpinner("The simulator is taking too long to initialize. Please try again later.");
+          return;
+        }
+      }
+
+    } catch (error) {
+      this.failSpinner("Error waiting for the simulator to be ready", error);
+      return;
     }
 
-  } catch (error) {
-    console.error(error);
+    if (resetDb) {
+      this.setSpinnerText("Resetting database...");
+      await this.simulatorService.cleanDatabase();
+    }
+
+    if (resetValidators) {
+      this.setSpinnerText("Initializing validators...");
+      try {
+        await this.simulatorService.deleteAllValidators();
+
+        const questions: DistinctQuestion[] = [
+          {
+            type: "checkbox",
+            name: "selectedLlmProviders",
+            message: "Select which LLM providers do you want to use:",
+            choices: this.simulatorService.getAiProvidersOptions(false),
+            validate: (answer) => (answer.length < 1 ? "You must choose at least one option." : true),
+          },
+        ];
+
+        const llmProvidersAnswer = await inquirer.prompt(questions);
+        await this.simulatorService.createRandomValidators(numValidators, llmProvidersAnswer.selectedLlmProviders);
+      } catch (error) {
+        this.failSpinner("Unable to initialize the validators", error);
+        return;
+      }
+    }
+
+    let successMessage = "GenLayer simulator initialized successfully! ";
+    successMessage += headless ? "" : `Go to ${this.simulatorService.getFrontendUrl()} in your browser to access it.`;
+    this.succeedSpinner(successMessage);
+
+    if (!headless) {
+      try {
+        this.startSpinner("Opening frontend...");
+        await this.simulatorService.openFrontend();
+        this.succeedSpinner("Frontend opened successfully");
+      } catch (error) {
+        this.failSpinner("Error opening the frontend", error);
+      }
+    }
   }
 }
