@@ -1,72 +1,71 @@
 import Docker from "dockerode";
 import { rpcClient } from "../../lib/clients/jsonRpcClient";
+import { BaseAction } from "../../lib/actions/BaseAction";
 
-export class OllamaAction {
+export class OllamaAction extends BaseAction {
   private docker: Docker;
 
   constructor() {
+    super();
     this.docker = new Docker();
   }
 
   async updateModel(modelName: string) {
-    const providersAndModels = await rpcClient.request({
-      method: "sim_getProvidersAndModels",
-      params: [],
-    });
+    try {
+      this.startSpinner(`Updating model "${modelName}"...`);
 
-    const existingOllamaProvider = providersAndModels.result.find(
-      (entry: any) => entry.plugin === "ollama"
-    );
+      if(!modelName){
+        modelName = this.getConfig().defaultOllamaModel;
+      }
 
-    if (!existingOllamaProvider) {
-      throw new Error("No existing 'ollama' provider found. Unable to add/update a model.");
-    }
-
-    await this.executeModelCommand(
-      "pull",
-      modelName,
-      `Model "${modelName}" updated successfully`
-    );
-
-    const existingModel = providersAndModels.result.some(
-      (entry: any) =>
-        entry.plugin === "ollama" && entry.model === modelName
-    );
-
-    if (!existingModel) {
-      console.log(`Model "${modelName}" not found in Provider Presets. Adding...`);
-
-      const newModelConfig = {
-        config: existingOllamaProvider.config,
-        model: modelName,
-        plugin: "ollama",
-        plugin_config: existingOllamaProvider.plugin_config,
-        provider: "ollama",
-      };
-
-      await rpcClient.request({
-        method: "sim_addProvider",
-        params: [newModelConfig],
+      const providersAndModels = await rpcClient.request({
+        method: "sim_getProvidersAndModels",
+        params: [],
       });
 
-      console.log(`Model "${modelName}" added successfully.`);
+      const existingOllamaProvider = providersAndModels.result.find(
+        (entry: any) => entry.plugin === "ollama"
+      );
+
+      if (!existingOllamaProvider) {
+        return this.failSpinner("No existing 'ollama' provider found. Unable to add/update a model.");
+      }
+
+      await this.executeModelCommand("pull", modelName, `Model "${modelName}" updated successfully`);
+
+      const existingModel = providersAndModels.result.some(
+        (entry: any) => entry.plugin === "ollama" && entry.model === modelName
+      );
+      if (!existingModel) {
+        this.startSpinner(`Adding model "${modelName}" to Provider Presets...`);
+
+        const newModelConfig = {
+          config: existingOllamaProvider.config,
+          model: modelName,
+          plugin: "ollama",
+          plugin_config: existingOllamaProvider.plugin_config,
+          provider: "ollama",
+        };
+
+        await rpcClient.request({
+          method: "sim_addProvider",
+          params: [newModelConfig],
+        });
+        this.succeedSpinner(`Model "${modelName}" added to Provider Presets successfully.`);
+      }
+    } catch (error) {
+      this.failSpinner(`Error updating model "${modelName}"`, error);
     }
   }
 
   async removeModel(modelName: string) {
-    await this.executeModelCommand(
-      "rm",
-      modelName,
-      `Model "${modelName}" removed successfully`
-    );
+    await this.executeModelCommand("rm", modelName, `Model "${modelName}" removed successfully`);
   }
 
-  private async executeModelCommand(
-    command: string,
-    modelName: string,
-    successMessage: string
-  ) {
+  private async executeModelCommand(command: string, modelName: string, successMessage: string) {
     try {
+      this.startSpinner(`Executing '${command}' command on model "${modelName}"...`);
+
       let success = false;
       const ollamaContainer = this.docker.getContainer("ollama");
       const exec = await ollamaContainer.exec({
@@ -74,12 +73,13 @@ export class OllamaAction {
         AttachStdout: true,
         AttachStderr: true,
       });
-      const stream = await exec.start({Detach: false, Tty: false});
+      const stream = await exec.start({ Detach: false, Tty: false });
 
       stream.on("data", (chunk: any) => {
-        const chunkStr = chunk.toString();
-        console.log(chunkStr);
-        if (chunkStr.includes("success") || chunkStr.includes("deleted")) {
+        const output = chunk.toString();
+        this.setSpinnerText(output.trim());
+
+        if (output.includes("success") || output.includes("deleted")) {
           success = true;
         }
       });
@@ -87,17 +87,17 @@ export class OllamaAction {
       await new Promise<void>((resolve, reject) => {
         stream.on("end", () => {
           if (success) {
+            this.succeedSpinner(successMessage);
             resolve();
           } else {
+            this.failSpinner(`Failed to execute '${command}' on model "${modelName}".`);
             reject('internal error');
           }
         });
         stream.on("error", reject);
       });
-
-      console.log(successMessage);
-    }catch (error) {
-      console.error(`Error executing command "${command}" on model "${modelName}":`, error);
+    } catch (error) {
+      this.failSpinner(`Error executing command "${command}" on model "${modelName}"`, error);
     }
   }
 }
