@@ -1,7 +1,6 @@
 import { describe, test, vi, beforeEach, afterEach, expect, Mock } from "vitest";
 import { OllamaAction } from "../../src/commands/update/ollama";
 import { rpcClient } from "../../src/lib/clients/jsonRpcClient";
-
 import Docker from "dockerode";
 
 vi.mock("dockerode");
@@ -37,6 +36,11 @@ describe("OllamaAction", () => {
     } as unknown as Docker.Container);
 
     Docker.prototype.getContainer = mockGetContainer;
+
+    vi.spyOn(ollamaAction as any, "startSpinner").mockImplementation(() => {});
+    vi.spyOn(ollamaAction as any, "setSpinnerText").mockImplementation(() => {});
+    vi.spyOn(ollamaAction as any, "succeedSpinner").mockImplementation(() => {});
+    vi.spyOn(ollamaAction as any, "failSpinner").mockImplementation(() => {});
   });
 
   afterEach(() => {
@@ -49,6 +53,7 @@ describe("OllamaAction", () => {
       config: { key: "value" },
       plugin_config: { pluginKey: "pluginValue" },
     };
+
     vi.mocked(rpcClient.request).mockResolvedValueOnce({
       result: [mockProvider],
     });
@@ -62,10 +67,9 @@ describe("OllamaAction", () => {
       }
     });
 
-    console.log = vi.fn();
-
     await ollamaAction.updateModel("mocked_model");
 
+    expect(ollamaAction["startSpinner"]).toHaveBeenCalledWith(`Updating model "mocked_model"...`);
     expect(mockGetContainer).toHaveBeenCalledWith("ollama");
     expect(mockExec).toHaveBeenCalledWith({
       Cmd: ["ollama", "pull", "mocked_model"],
@@ -73,10 +77,8 @@ describe("OllamaAction", () => {
       AttachStderr: true,
     });
     expect(mockStart).toHaveBeenCalledWith({ Detach: false, Tty: false });
-    expect(mockStream.on).toHaveBeenCalledWith("data", expect.any(Function));
-    expect(mockStream.on).toHaveBeenCalledWith("end", expect.any(Function));
-    expect(console.log).toHaveBeenCalledWith("Mocked output success");
-    expect(console.log).toHaveBeenCalledWith('Model "mocked_model" updated successfully');
+    expect(ollamaAction["setSpinnerText"]).toHaveBeenCalledWith("Mocked output success");
+    expect(ollamaAction["succeedSpinner"]).toHaveBeenCalledWith(`Model "mocked_model" updated successfully`);
   });
 
   test("should remove the model using 'rm'", async () => {
@@ -89,10 +91,9 @@ describe("OllamaAction", () => {
       }
     });
 
-    console.log = vi.fn();
-
     await ollamaAction.removeModel("mocked_model");
 
+    expect(ollamaAction["startSpinner"]).toHaveBeenCalledWith(`Executing 'rm' command on model "mocked_model"...`);
     expect(mockGetContainer).toHaveBeenCalledWith("ollama");
     expect(mockExec).toHaveBeenCalledWith({
       Cmd: ["ollama", "rm", "mocked_model"],
@@ -100,10 +101,8 @@ describe("OllamaAction", () => {
       AttachStderr: true,
     });
     expect(mockStart).toHaveBeenCalledWith({ Detach: false, Tty: false });
-    expect(mockStream.on).toHaveBeenCalledWith("data", expect.any(Function));
-    expect(mockStream.on).toHaveBeenCalledWith("end", expect.any(Function));
-    expect(console.log).toHaveBeenCalledWith("Mocked output success");
-    expect(console.log).toHaveBeenCalledWith('Model "mocked_model" removed successfully');
+    expect(ollamaAction["setSpinnerText"]).toHaveBeenCalledWith("Mocked output success");
+    expect(ollamaAction["succeedSpinner"]).toHaveBeenCalledWith(`Model "mocked_model" removed successfully`);
   });
 
   test("should log an error if an exception occurs during 'pull'", async () => {
@@ -112,48 +111,26 @@ describe("OllamaAction", () => {
       config: { key: "value" },
       plugin_config: { pluginKey: "pluginValue" },
     };
+
     vi.mocked(rpcClient.request).mockResolvedValueOnce({
       result: [mockProvider],
     });
 
     const error = new Error("Mocked error");
-    mockGetContainer.mockReturnValueOnce(
-      {
-        exec: () => {
-          throw new Error("Mocked error");
-        }
-      }
-    );
-    console.error = vi.fn();
+    mockExec.mockRejectedValue(error);
 
     await ollamaAction.updateModel("mocked_model");
 
-    expect(mockGetContainer).toHaveBeenCalledWith("ollama");
-    expect(console.error).toHaveBeenCalledWith(
-      'Error executing command "pull" on model "mocked_model":',
-      error
-    );
+    expect(ollamaAction["failSpinner"]).toHaveBeenCalledWith(`Error executing command "pull" on model "mocked_model"`, error);
   });
 
   test("should log an error if an exception occurs during 'rm'", async () => {
     const error = new Error("Mocked error");
-    mockGetContainer.mockReturnValueOnce(
-      {
-        exec: () => {
-          throw new Error("Mocked error");
-        }
-      }
-    );
-
-    console.error = vi.fn();
+    mockExec.mockRejectedValue(error);
 
     await ollamaAction.removeModel("mocked_model");
 
-    expect(mockGetContainer).toHaveBeenCalledWith("ollama");
-    expect(console.error).toHaveBeenCalledWith(
-      'Error executing command "rm" on model "mocked_model":',
-      error
-    );
+    expect(ollamaAction["failSpinner"]).toHaveBeenCalledWith(`Error executing command "rm" on model "mocked_model"`, error);
   });
 
   test("should throw an error if no 'ollama' provider exists during updateModel", async () => {
@@ -161,21 +138,14 @@ describe("OllamaAction", () => {
       result: [],
     });
 
-    const modelName = "mocked_model";
+    await ollamaAction.updateModel("mocked_model");
 
-    await expect(ollamaAction.updateModel(modelName)).rejects.toThrowError(
+    expect(ollamaAction["failSpinner"]).toHaveBeenCalledWith(
       "No existing 'ollama' provider found. Unable to add/update a model."
     );
-
-    expect(rpcClient.request).toHaveBeenCalledWith({
-      method: "sim_getProvidersAndModels",
-      params: [],
-    });
   });
 
   test("should reject with an error if success is not set to true", async () => {
-    console.error = vi.fn();
-
     const mockProvider = {
       plugin: "ollama",
       config: { key: "value" },
@@ -195,14 +165,29 @@ describe("OllamaAction", () => {
       }
     });
 
-    console.log = vi.fn();
-    console.error = vi.fn();
+    await ollamaAction.updateModel("mocked_model");
+
+    expect(ollamaAction["failSpinner"]).toHaveBeenCalledWith(
+      `Failed to execute 'pull' on model "mocked_model".`
+    );
+  });
+
+  test("should log an error if an exception occurs inside updateModel", async () => {
+    const mockError = new Error("Mocked error");
+
+    vi.mocked(rpcClient.request).mockRejectedValue(mockError);
 
     await ollamaAction.updateModel("mocked_model");
 
-    expect(console.error).toHaveBeenCalledWith(
-      'Error executing command "pull" on model "mocked_model":', 'internal error'
-    );
+    expect(ollamaAction["failSpinner"]).toHaveBeenCalledWith(`Error updating model "mocked_model"`, mockError);
+  });
+
+  test("should call get config if modelName is empty", async () => {
+    const defaultModel = "default_model";
+    vi.spyOn(ollamaAction as any, "getConfig").mockReturnValue({ defaultOllamaModel: defaultModel });
+
+    await ollamaAction.updateModel("");
+    expect(ollamaAction.getConfig).toHaveBeenCalledTimes(1);
   });
 
 });
