@@ -1,175 +1,174 @@
 import { describe, beforeEach, afterEach, test, expect, vi, Mock } from "vitest";
 import inquirer from "inquirer";
-import { startAction, StartActionOptions } from "../../src/commands/general/start";
-import { ISimulatorService } from "../../src/lib/interfaces/ISimulatorService";
+import { StartAction, StartActionOptions } from "../../src/commands/general/start";
+import { SimulatorService } from "../../src/lib/services/simulator";
 
-describe("startAction - Additional Tests", () => {
-  let simulatorService: ISimulatorService;
-  let logSpy: ReturnType<typeof vi.spyOn>;
-  let errorSpy: ReturnType<typeof vi.spyOn>;
-  let promptSpy: ReturnType<any>;
+vi.mock("../../src/lib/services/simulator");
+vi.mock("inquirer");
 
-  const defaultOptions: StartActionOptions = {
-    resetValidators: false,
-    numValidators: 5,
-    headless: false,
-    resetDb: false
-  };
+describe("StartAction", () => {
+  let startAction: StartAction;
+  let mockSimulatorService: SimulatorService;
 
   beforeEach(() => {
-    logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
-    errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-    promptSpy = vi.spyOn(inquirer, "prompt");
+    vi.clearAllMocks();
 
-    simulatorService = {
-      updateSimulator: vi.fn().mockResolvedValue(undefined),
-      runSimulator: vi.fn().mockResolvedValue(undefined),
-      waitForSimulatorToBeReady: vi.fn().mockResolvedValue({ initialized: true }),
-      deleteAllValidators: vi.fn().mockResolvedValue(undefined),
-      createRandomValidators: vi.fn().mockResolvedValue(undefined),
-      openFrontend: vi.fn().mockResolvedValue(undefined),
-      setSimulatorLocation: vi.fn().mockResolvedValue(undefined),
-      setComposeOptions: vi.fn(),
-      checkCliVersion: vi.fn(),
-      getAiProvidersOptions: vi.fn(() => [
-        { name: "Provider A", value: "providerA" },
-        { name: "Provider B", value: "providerB" },
-      ]),
-      getFrontendUrl: vi.fn(() => "http://localhost:8080"),
-      cleanDatabase: vi.fn().mockResolvedValue(undefined),
-    } as unknown as ISimulatorService;
+    mockSimulatorService = new SimulatorService();
+    startAction = new StartAction();
+    startAction["simulatorService"] = mockSimulatorService;
+
+    mockSimulatorService.waitForSimulatorToBeReady = vi.fn().mockResolvedValue({ initialized: true });
+
+    vi.spyOn(startAction as any, "startSpinner").mockImplementation(() => {});
+    vi.spyOn(startAction as any, "setSpinnerText").mockImplementation(() => {});
+    vi.spyOn(startAction as any, "succeedSpinner").mockImplementation(() => {});
+    vi.spyOn(startAction as any, "failSpinner").mockImplementation(() => {});
   });
 
   afterEach(() => {
     vi.restoreAllMocks();
   });
 
-  test("runs successfully with default options and keeps existing validators", async () => {
-    await startAction(defaultOptions, simulatorService);
+  const defaultOptions: StartActionOptions = {
+    resetValidators: false,
+    numValidators: 5,
+    headless: false,
+    resetDb: false,
+  };
 
-    expect(simulatorService.runSimulator).toHaveBeenCalled();
-    expect(simulatorService.waitForSimulatorToBeReady).toHaveBeenCalled();
+  test("should start the simulator successfully", async () => {
+    mockSimulatorService.checkCliVersion = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.runSimulator = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.getFrontendUrl = vi.fn().mockReturnValue("http://localhost:8080");
 
-    expect(logSpy).toHaveBeenCalledWith("Starting GenLayer simulator keeping the existing validators");
-    expect(logSpy).toHaveBeenCalledWith("Simulator is running!");
-    expect(logSpy).toHaveBeenCalledWith("GenLayer simulator initialized successfully! Go to http://localhost:8080 in your browser to access it.");
+    await startAction.execute(defaultOptions);
 
-    expect(simulatorService.openFrontend).toHaveBeenCalled();
+    expect(startAction["startSpinner"]).toHaveBeenCalledWith("Checking CLI version...");
+    expect(mockSimulatorService.checkCliVersion).toHaveBeenCalled();
+
+    expect(startAction["setSpinnerText"]).toHaveBeenCalledWith("Starting GenLayer Localnet (keeping the existing validators)...");
+    expect(mockSimulatorService.runSimulator).toHaveBeenCalled();
+
+    expect(startAction["setSpinnerText"]).toHaveBeenCalledWith("Waiting for the simulator to be ready...");
+    expect(mockSimulatorService.waitForSimulatorToBeReady).toHaveBeenCalled();
+
+    expect(startAction["succeedSpinner"]).toHaveBeenCalledWith("GenLayer simulator initialized successfully! Go to http://localhost:8080 in your browser to access it.");
   });
 
-  test("runs successfully with custom options and keeps existing validators", async () => {
-    await startAction({...defaultOptions, headless: true, resetDb: true}, simulatorService);
-
-    expect(simulatorService.runSimulator).toHaveBeenCalled();
-    expect(simulatorService.waitForSimulatorToBeReady).toHaveBeenCalled();
-
-    expect(logSpy).toHaveBeenCalledWith("Starting GenLayer simulator keeping the existing validators");
-
-    expect(logSpy).toHaveBeenCalledWith("Simulator is running!");
-    expect(logSpy).toHaveBeenCalledWith("GenLayer simulator initialized successfully! ");
-  });
-
-
-  test("logs error and stops if runSimulator fails", async () => {
+  test("should fail when simulator fails to start", async () => {
     const errorMsg = new Error("runSimulator error");
-    (simulatorService.runSimulator as Mock).mockRejectedValueOnce(errorMsg);
+    (mockSimulatorService.runSimulator as Mock).mockRejectedValueOnce(errorMsg);
 
-    await startAction(defaultOptions, simulatorService);
+    await startAction.execute(defaultOptions);
 
-    expect(errorSpy).toHaveBeenCalledWith(errorMsg);
-    expect(simulatorService.waitForSimulatorToBeReady).not.toHaveBeenCalled();
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("Error starting the simulator", errorMsg);
   });
 
-  test("handles resetfValidators correctly by deleting and creating new validators", async () => {
-    promptSpy.mockResolvedValueOnce({ selectedLlmProviders: ["providerA"] });
-    const optionsWithReset: StartActionOptions = { ...defaultOptions, resetValidators: true };
+  test("should fail when waiting for simulator initialization times out", async () => {
+    (mockSimulatorService.waitForSimulatorToBeReady as Mock).mockResolvedValue({ initialized: false, errorCode: "TIMEOUT" });
 
-    await startAction(optionsWithReset, simulatorService);
+    await startAction.execute(defaultOptions);
 
-    expect(simulatorService.deleteAllValidators).toHaveBeenCalled();
-    expect(simulatorService.createRandomValidators).toHaveBeenCalledWith(5, ["providerA"]);
-    expect(logSpy).toHaveBeenCalledWith("New random validators successfully created...");
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("The simulator is taking too long to initialize. Please try again later.");
   });
 
-  test("logs error if deleteAllValidators fails when resetValidators is true", async () => {
-    const errorMsg = new Error("deleteAllValidators error");
-    (simulatorService.deleteAllValidators as Mock).mockRejectedValueOnce(errorMsg);
-    const optionsWithReset: StartActionOptions = { ...defaultOptions, resetValidators: true };
+  test("should reset the database if resetDb is true", async () => {
+    const options: StartActionOptions = { ...defaultOptions, resetDb: true };
 
-    await startAction(optionsWithReset, simulatorService);
+    mockSimulatorService.cleanDatabase = vi.fn().mockResolvedValue(undefined);
 
-    expect(errorSpy).toHaveBeenCalledWith("Unable to initialize the validators.");
-    expect(errorSpy).toHaveBeenCalledWith(errorMsg);
-    expect(simulatorService.createRandomValidators).not.toHaveBeenCalled();
+    await startAction.execute(options);
+
+    expect(startAction["setSpinnerText"]).toHaveBeenCalledWith("Resetting database...");
+    expect(mockSimulatorService.cleanDatabase).toHaveBeenCalled();
   });
 
-  test("prompts for LLM providers and validates at least one option is selected", async () => {
-    const aiProviders = simulatorService.getAiProvidersOptions(false);
-    expect(aiProviders).toEqual([
-      { name: "Provider A", value: "providerA" },
-      { name: "Provider B", value: "providerB" },
-    ]);
+  test("should initialize validators when resetValidators is true", async () => {
+    const options: StartActionOptions = { ...defaultOptions, resetValidators: true };
 
-    promptSpy.mockImplementation(async (questions: any) => {
-      const validateFunction = questions[0].validate;
+    mockSimulatorService.deleteAllValidators = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.createRandomValidators = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.getAiProvidersOptions = vi.fn().mockReturnValue(["Provider1", "Provider2"]);
 
-      expect(validateFunction([])).toBe("You must choose at least one option.");
-      expect(validateFunction(["providerA"])).toBe(true);
+    vi.mocked(inquirer.prompt).mockResolvedValue({ selectedLlmProviders: ["Provider1"] });
 
-      return { selectedLlmProviders: ["providerA"] };
-    });
+    await startAction.execute(options);
 
-    const optionsWithReset: StartActionOptions = { ...defaultOptions, resetValidators: true };
-    await startAction(optionsWithReset, simulatorService);
-
-    expect(promptSpy).toHaveBeenCalled();
-    expect(simulatorService.createRandomValidators).toHaveBeenCalledWith(5, ["providerA"]);
+    expect(startAction["setSpinnerText"]).toHaveBeenCalledWith("Initializing validators...");
+    expect(mockSimulatorService.deleteAllValidators).toHaveBeenCalled();
+    expect(mockSimulatorService.createRandomValidators).toHaveBeenCalledWith(5, ["Provider1"]);
   });
 
-  test("logs specific message if waitForSimulatorToBeReady returns TIMEOUT errorCode", async () => {
-    (simulatorService.waitForSimulatorToBeReady as Mock).mockResolvedValue({
+  test("should fail when initializing validators fails", async () => {
+    const options: StartActionOptions = { ...defaultOptions, resetValidators: true };
+
+    mockSimulatorService.deleteAllValidators = vi.fn().mockRejectedValue(new Error("Failed to delete validators"));
+
+    await startAction.execute(options);
+
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("Unable to initialize the validators", expect.any(Error));
+  });
+
+  test("should open frontend when not in headless mode", async () => {
+    mockSimulatorService.checkCliVersion = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.runSimulator = vi.fn().mockResolvedValue(undefined);
+    mockSimulatorService.getFrontendUrl = vi.fn().mockReturnValue("http://localhost:8080");
+    mockSimulatorService.openFrontend = vi.fn().mockResolvedValue(undefined);
+
+    await startAction.execute(defaultOptions);
+
+    expect(startAction["startSpinner"]).toHaveBeenCalledWith("Opening frontend...");
+    expect(mockSimulatorService.openFrontend).toHaveBeenCalled();
+    expect(startAction["succeedSpinner"]).toHaveBeenCalledWith("Frontend opened successfully");
+  });
+
+  test("should handle errors when opening frontend", async () => {
+    const errorMsg = new Error("Failed to open frontend");
+    (mockSimulatorService.openFrontend as Mock).mockRejectedValueOnce(errorMsg);
+
+    await startAction.execute(defaultOptions);
+
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("Error opening the frontend", errorMsg);
+  });
+
+  test("should log specific message if waitForSimulatorToBeReady returns TIMEOUT errorCode", async () => {
+    (mockSimulatorService.waitForSimulatorToBeReady as Mock).mockResolvedValue({
       initialized: false,
       errorCode: "TIMEOUT",
       errorMessage: "Initialization timed out",
     });
 
-    await startAction(defaultOptions, simulatorService);
+    await startAction.execute(defaultOptions);
 
-    expect(errorSpy).toHaveBeenCalledWith(
-      "The simulator is taking too long to initialize. Please try again after the simulator is ready."
-    );
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("The simulator is taking too long to initialize. Please try again later.");
   });
 
-  test("logs error message if simulator fails to initialize with ERROR code", async () => {
-    (simulatorService.waitForSimulatorToBeReady as Mock).mockResolvedValue({
+  test("should log error message if simulator fails to initialize with ERROR code", async () => {
+    (mockSimulatorService.waitForSimulatorToBeReady as Mock).mockResolvedValue({
       initialized: false,
       errorCode: "ERROR",
       errorMessage: "Initialization failed",
     });
 
-    await startAction(defaultOptions, simulatorService);
+    await startAction.execute(defaultOptions);
 
-    expect(logSpy).toHaveBeenCalledWith("Initialization failed");
-    expect(errorSpy).toHaveBeenCalledWith("Unable to initialize the GenLayer simulator. Please try again.");
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("Unable to initialize the GenLayer simulator.", "Initialization failed");
   });
 
   test("catches and logs error if waitForSimulatorToBeReady throws an exception", async () => {
     const errorMsg = new Error("Unexpected initialization error");
-    (simulatorService.waitForSimulatorToBeReady as Mock).mockRejectedValueOnce(errorMsg);
+    (mockSimulatorService.waitForSimulatorToBeReady as Mock).mockRejectedValueOnce(errorMsg);
 
-    await startAction(defaultOptions, simulatorService);
+    await startAction.execute(defaultOptions);
 
-    expect(errorSpy).toHaveBeenCalledWith(errorMsg);
+    expect(startAction["failSpinner"]).toHaveBeenCalledWith("Error waiting for the simulator to be ready", errorMsg);
   });
 
-  test("catches and logs error if openFrontend throws an exception", async () => {
-    const errorMsg = new Error("Failed to open frontend");
-    (simulatorService.openFrontend as Mock).mockImplementationOnce(() => {
-      throw errorMsg;
-    });
+  test("should not append frontend URL when in headless mode", async () => {
+    await startAction.execute({ ...defaultOptions, headless: true });
 
-    await startAction(defaultOptions, simulatorService);
-
-    expect(errorSpy).toHaveBeenCalledWith(errorMsg);
+    expect(startAction["succeedSpinner"]).toHaveBeenCalledWith(
+      "GenLayer simulator initialized successfully! "
+    );
   });
 });
