@@ -2,6 +2,8 @@ import { describe, test, vi, beforeEach, afterEach, expect } from "vitest";
 import { StopAction } from "../../src/commands/general/stop";
 import { SimulatorService } from "../../src/lib/services/simulator";
 import { ISimulatorService } from "../../src/lib/interfaces/ISimulatorService";
+import chalk from "chalk";
+
 import inquirer from "inquirer";
 
 vi.mock("../../src/lib/services/simulator");
@@ -22,6 +24,10 @@ describe("StopAction", () => {
 
     stopAction = new StopAction();
     (stopAction as any).simulatorService = mockSimulatorService;
+
+    vi.spyOn(stopAction as any, "startSpinner").mockImplementation(() => {});
+    vi.spyOn(stopAction as any, "succeedSpinner").mockImplementation(() => {});
+    vi.spyOn(stopAction as any, "failSpinner").mockImplementation(() => {});
   });
 
   afterEach(() => {
@@ -37,22 +43,35 @@ describe("StopAction", () => {
       {
         type: "confirm",
         name: "confirmAction",
-        message: "Are you sure you want to stop all running GenLayer containers? This will halt all active processes.",
+        message: chalk.yellow("Are you sure you want to stop all running GenLayer containers? This will halt all active processes."),
         default: true,
       },
     ]);
     expect(mockSimulatorService.stopDockerContainers).toHaveBeenCalled();
+    expect(stopAction["succeedSpinner"]).toHaveBeenCalledWith(
+      "All running GenLayer containers have been successfully stopped."
+    );
   });
 
   test("should abort if user cancels", async () => {
     vi.mocked(inquirer.prompt).mockResolvedValue({ confirmAction: false });
 
-    console.log = vi.fn();
-
     await stopAction.stop();
 
     expect(inquirer.prompt).toHaveBeenCalled();
-    expect(console.log).toHaveBeenCalledWith("Operation aborted!");
     expect(mockSimulatorService.stopDockerContainers).not.toHaveBeenCalled();
+  });
+
+  test("should handle errors and call failSpinner", async () => {
+    vi.mocked(inquirer.prompt).mockResolvedValue({ confirmAction: true });
+    const error = new Error("Test Error");
+    mockSimulatorService.stopDockerContainers = vi.fn().mockRejectedValue(error);
+
+    await stopAction.stop();
+
+    expect(stopAction["failSpinner"]).toHaveBeenCalledWith(
+      "An error occurred while stopping the containers.",
+      error
+    );
   });
 });
