@@ -2,10 +2,8 @@ import { Command } from "commander";
 import { vi, describe, beforeEach, afterEach, test, expect } from "vitest";
 import { initializeUpdateCommands } from "../../src/commands/update";
 import { OllamaAction } from "../../src/commands/update/ollama";
-import { ConfigFileManager } from "../../src/lib/config/ConfigFileManager";
 
 vi.mock("../../src/commands/update/ollama");
-vi.mock("../../src/lib/config/ConfigFileManager");
 
 describe("ollama command", () => {
   let program: Command;
@@ -15,7 +13,6 @@ describe("ollama command", () => {
     initializeUpdateCommands(program);
 
     const mockConfig = { defaultOllamaModel: "default-model" };
-    vi.mocked(ConfigFileManager.prototype.getConfig).mockReturnValue(mockConfig);
   });
 
   afterEach(() => {
@@ -31,7 +28,7 @@ describe("ollama command", () => {
   test("OllamaAction.updateModel is called with default model", async () => {
     program.parse(["node", "test", "update", "ollama"]);
     expect(OllamaAction).toHaveBeenCalledTimes(1);
-    expect(OllamaAction.prototype.updateModel).toHaveBeenCalledWith("default-model");
+    expect(OllamaAction.prototype.updateModel).toHaveBeenCalledWith("");
   });
 
   test("OllamaAction.removeModel is called with model option", async () => {
@@ -43,6 +40,6 @@ describe("ollama command", () => {
   test("OllamaAction.removeModel is called with default model", async () => {
     program.parse(["node", "test", "update", "ollama", "--remove"]);
     expect(OllamaAction).toHaveBeenCalledTimes(1);
-    expect(OllamaAction.prototype.removeModel).toHaveBeenCalledWith("default-model");
+    expect(OllamaAction.prototype.removeModel).toHaveBeenCalledWith("");
   });
 });
